main.go
// Declare the main package
package main

// Import necessary packages
import (
	"bufio" // Scanner to read text
	"fmt"   // Formatting package for input/output
	"os"    // Operating system interface, for files and command-line arguments
)

// Entry point of the Go program
func main() {
	// Check for the correct number of command-line arguments
	if len(os.Args) != 3 {
		// If incorrect, print usage to console
		fmt.Printf("Usage: %s <inputfile> <outputfile>\n", os.Args[0])
		// Exit the program with a status code of 1 (indicating an error)
		os.Exit(1)
	}

	// Assign command-line arguments to variables
	inputFileName := os.Args[1]  // Input file name
	outputFileName := os.Args[2] // Output file name

	// Attempt to open input file for reading
	inputFile, err := os.Open(inputFileName)
	// Check for error while opening file
	if err != nil {
		// If there's an error, display it and exit
		fmt.Printf("Error opening input file: %v\n", err)
		os.Exit(1)
	}
	// Schedule the input file to be closed when the function exits
	defer inputFile.Close()

	// Create a new Scanner to read from inputFile
	scanner := bufio.NewScanner(inputFile)
	// Declare an empty string to accumulate text
	var text string
	// Loop over all scan events
	for scanner.Scan() {
		// Concatenate each line of text and a newline character
		text += scanner.Text() + "\n"
	}

	// Check for any errors that occurred during scanning
	if err := scanner.Err(); err != nil {
		// Print any scanning errors
		fmt.Printf("Error reading input file: %v\n", err)
		// Exit with an error status
		os.Exit(1)
	}

	// Process the text to apply transformations
	processedText := ProcessText(text)

	// Attempt to create the output file
	outputFile, err := os.Create(outputFileName)
	// Check for errors while creating the file
	if err != nil {
		// Print an error message and exit if a file create error occurs
		fmt.Printf("Error creating output file: %v\n", err)
		os.Exit(1)
	}
	// Schedule the output file to be closed when the function exits
	defer outputFile.Close()

	// Create a buffered writer for the output file
	writer := bufio.NewWriter(outputFile)
	// Write the processed text to the output file buffer
	_, err = writer.WriteString(processedText)
	// Check for an error while writing
	if err != nil {
		// Print write error and exit if one occurs
		fmt.Printf("Error writing to output file: %v\n", err)
		os.Exit(1)
	}
	// Flush the buffer to ensure all content is written to the file
	writer.Flush()
}

// ProcessText performs all the text transformations.
func ProcessText(text string) string {
	// Here we would call functions for processing the text.
	// These functions should be implemented to perform the specified modifications.

	// Replace hexadecimal and binary numbers with their decimal equivalents
	text = ReplaceNums(text)
	// Apply casing rules: uppercase, lowercase, and capitalized
	text = ApplyCasing(text)
	// Correct punctuation spacing and usage
	text = FormatPunctuation(text)
	// Correct single quotes
	text = FormatQuotes(text)
	// Correct the usage of indefinite articles 'a' and 'an'
	text = AdjustArticle(text)

	// Return the processed text
	return text
}

// Note: Functions such as ReplaceNums(), ApplyCasing(), FormatPunctuation(), FormatQuotes(),
// and AdjustArticle() would need to be defined with their own implementation logic.
